app/requirements.txt
fastapi==0.110.1
uvicorn==0.15.0
boto3==1.34.76
psycopg2-binary==2.9.9
pydantic-settings==2.2.1
transformers==4.39.3
python-dotenv==1.0.1
SQLAlchemy==2.0.25
scipy==1.13.0
torch==2.2.2
